Added License and changed packages
