+++ camel-example-spring-boot-load-balancer-eip/README.adoc
@@ -2,7 +2,7 @@
 
 This example shows several examples of Load Balancer EIP with Apache Camel application using Spring Boot.
 
-1. Round-robin - The exchanges are selected from in a round robin fashion. This is a well known and classic policy, which spreads the load evenly.
+1. Round-robin - The exchanges are selected in a round robin fashion. This is a well known and classic policy, which spreads the load evenly.
 2. Random load - A random endpoint is selected for each exchange.
 3. Sticky - Sticky load balancing using an Expression to calculate a correlation key to perform the sticky load balancing; rather like jsessionid in the web or JMSXGroupID in JMS.
 4. Topic - Topic which sends to all destinations (rather like JMS Topics)
