--- camel-example-spring-boot-load-balancer-eip/src/main/java/sample/camel/MyException.java
@@ -1,11 +1,0 @@
-package sample.camel;
-
-public class MyException extends RuntimeException {
-
-	MyException() {
-	}
-
-	MyException(String message) {
-		super(message);
-	}
-}
